streamlitapp/requirements.txt
streamlit
pandas
numpy
opencv-python
Pillow
scipy
plotly






#streamlit run app.py